+++ com.unity.project-auditor/Editor/API/ReportItem.cs
@@ -43,7 +43,7 @@ namespace Unity.ProjectAuditor.Editor
         /// An unique identifier for the issue descriptor (read-only).
         /// </summary>
         /// <remarks>
-        /// Project Reports can contain two different types of ReportItem:
+        /// Reports can contain two different types of ReportItem:
         /// - Issues, which indicate a potential problem which should be investigated and possibly fixed: for example, a texture with its Read/Write Enabled checkbox ticked.
         /// - Insights, for informational purposes: for example, general information about a texture in the project.
         ///
@@ -79,7 +79,7 @@ namespace Unity.ProjectAuditor.Editor
 
         /// <summary>
         /// Custom properties.
-        /// See the "moduleMetadata" section of an exported Project Report JSON file for information on the formats and
+        /// See the "moduleMetadata" section of an exported Report JSON file for information on the formats and
         /// meanings of the custom properties for each IssueCategory.
         /// </summary>
         [JsonProperty("properties")]