+++ com.unity.project-auditor/Editor/Core/DescriptorLibrary.cs
@@ -52,7 +52,7 @@ namespace Unity.ProjectAuditor.Editor.Core
         {
             // update list from dictionary
 
-            // TODO: Serialization is needed to survive domain reload, and when writing a ProjectReport out to file.
+            // TODO: Serialization is needed to survive domain reload, and when writing a Report out to file.
             // In both cases the list only really needs to contain the Descriptors that correspond to ProjectIssues
             // actually found in the report, so if we had the report object we could potentially do some filtering here.
             m_SerializedDescriptors = s_Descriptors.Values.ToList();