+++ com.unity.project-auditor/Editor/UI/Framework/ViewManager.cs
@@ -18,13 +18,13 @@ namespace Unity.ProjectAuditor.Editor.UI.Framework
             }
         }
 
-        ProjectReport m_Report;
+        Report m_Report;
         AnalysisView[] m_Views;
 
         [SerializeField] IssueCategory[] m_Categories;
         [SerializeField] int m_ActiveViewIndex;
 
-        public ProjectReport Report => m_Report;
+        public Report Report => m_Report;
 
         public int NumViews => m_Views != null ? m_Views.Length : 0;
 
@@ -193,13 +193,13 @@ namespace Unity.ProjectAuditor.Editor.UI.Framework
             }
         }
 
-        public void OnAnalysisCompleted(ProjectReport report)
+        public void OnAnalysisCompleted(Report report)
         {
             m_Report = report;
             MarkViewColumnWidthsAsDirty();
         }
 
-        public void OnAnalysisRestored(ProjectReport report)
+        public void OnAnalysisRestored(Report report)
         {
             AddIssues(report.GetAllIssues());
             m_Report = report;